Jim Ahlstrom patch: cutoff should be a long for machines with 16-bit ints.  (In theory, other variables should be widened to long as well, but this won't ever be needed, since the len of a list is still an int.)

+++ Objects/listobject.c
@@ -877,7 +877,7 @@ struct SamplesortStackNode {
    is undesirable, so cutoff values are canned in the "cutoff" table
    below:  cutoff[i] is the smallest N such that k == CUTOFFBASE + i. */
 #define CUTOFFBASE 4
-static int cutoff[] = {
+static long cutoff[] = {
 	43,        /* smallest N such that k == 4 */
 	106,       /* etc */
 	250,
